rename src/effects.js -> src/utils.js
